compiler: move VerifyPhase to jdk.graal.compiler.test
