+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyAArch64RegisterUsages.java
@@ -30,7 +30,6 @@ import java.util.List;
 import jdk.graal.compiler.nodes.StructuredGraph;
 import jdk.graal.compiler.nodes.java.LoadFieldNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.vm.ci.aarch64.AArch64;
 import jdk.vm.ci.code.Register;
 import jdk.vm.ci.meta.JavaField;
