+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyEndlessLoops.java
@@ -39,7 +39,6 @@ import jdk.graal.compiler.nodes.loop.Loop;
 import jdk.graal.compiler.nodes.loop.LoopsData;
 import jdk.graal.compiler.nodes.memory.MemoryKill;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.graal.compiler.phases.schedule.SchedulePhase;
 
 /**
