+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyInstanceOfUsage.java
@@ -30,8 +30,6 @@ import jdk.graal.compiler.lir.StandardOp.ValueMoveOp;
 import jdk.graal.compiler.nodes.StructuredGraph;
 import jdk.graal.compiler.nodes.java.InstanceOfNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
-import jdk.graal.compiler.phases.VerifyPhase;
-
 import jdk.vm.ci.meta.MetaAccessProvider;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
