+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyMathAbs.java
@@ -28,7 +28,6 @@ import jdk.graal.compiler.core.common.NumUtil;
 import jdk.graal.compiler.nodes.StructuredGraph;
 import jdk.graal.compiler.nodes.java.MethodCallTargetNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.vm.ci.meta.MetaAccessProvider;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
