+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyPhaseNoDirectRecursion.java
@@ -29,7 +29,6 @@ import jdk.graal.compiler.nodes.java.MethodCallTargetNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.graal.compiler.phases.BasePhase;
 import jdk.graal.compiler.phases.RecursivePhase;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
