+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyPluginFrameState.java
@@ -38,7 +38,6 @@ import jdk.graal.compiler.nodes.graphbuilderconf.InvocationPlugin;
 import jdk.graal.compiler.nodes.java.MethodCallTargetNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.graal.compiler.nodes.spi.Proxy;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.graal.compiler.replacements.InvocationPluginHelper;
 import jdk.graal.compiler.util.EconomicHashSet;
 import jdk.vm.ci.meta.MetaAccessProvider;
