+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyProfileMethodUsage.java
@@ -36,7 +36,6 @@ import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.graal.compiler.nodes.spi.ProfileProvider;
 import jdk.graal.compiler.nodes.spi.ResolvedJavaMethodProfileProvider;
 import jdk.graal.compiler.nodes.spi.StableProfileProvider;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
