+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyStatelessPhases.java
@@ -36,8 +36,6 @@ import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.graal.compiler.phases.BasePhase;
 import jdk.graal.compiler.phases.PhaseSuite;
 import jdk.graal.compiler.phases.SingleRunSubphase;
-import jdk.graal.compiler.phases.VerifyPhase;
-
 import jdk.vm.ci.meta.MetaAccessProvider;
 import jdk.vm.ci.meta.ResolvedJavaField;
