+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyStringCaseUsage.java
@@ -27,7 +27,6 @@ package jdk.graal.compiler.core.test;
 import jdk.graal.compiler.nodes.StructuredGraph;
 import jdk.graal.compiler.nodes.java.MethodCallTargetNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
