+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyUnsafeAccess.java
@@ -29,7 +29,6 @@ import java.lang.reflect.Field;
 import jdk.graal.compiler.nodes.StructuredGraph;
 import jdk.graal.compiler.nodes.java.MethodCallTargetNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
-import jdk.graal.compiler.phases.VerifyPhase;
 import jdk.vm.ci.meta.MetaAccessProvider;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
