+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/core/test/VerifyVirtualizableUsage.java
@@ -29,7 +29,6 @@ import jdk.graal.compiler.core.common.type.Stamp;
 import jdk.graal.compiler.graph.Graph;
 import jdk.graal.compiler.graph.Node;
 import jdk.graal.compiler.graph.NodeInputList;
-import jdk.graal.compiler.nodes.spi.VirtualizerTool;
 import jdk.graal.compiler.nodes.ConstantNode;
 import jdk.graal.compiler.nodes.NodeView;
 import jdk.graal.compiler.nodes.StructuredGraph;
@@ -37,8 +36,7 @@ import jdk.graal.compiler.nodes.ValueNode;
 import jdk.graal.compiler.nodes.java.MethodCallTargetNode;
 import jdk.graal.compiler.nodes.spi.CoreProviders;
 import jdk.graal.compiler.nodes.spi.Virtualizable;
-import jdk.graal.compiler.phases.VerifyPhase;
-
+import jdk.graal.compiler.nodes.spi.VirtualizerTool;
 import jdk.vm.ci.meta.ResolvedJavaMethod;
 import jdk.vm.ci.meta.ResolvedJavaType;
